Data Annotation/requirements.txt
imutils==0.5.3
lxml==4.5.2
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
PyQt5==5.15.0
PyQt5-sip==12.8.0


